Moved Message List Item to separate component
